add a cast function that suppresses -Wcast-function-type-strict (#100170)
add a cast function that suppresses -Wcast-function-type-strict

Summary:
These casts are a necessary evil due to the design of Python. Python
ultimately casts it back to the original type based on the flags
specified in the PyMethodDef.

Nevertheless, the new Clang flag -Wcast-function-type-strict breaks
with this.

Test Plan: Passes builds with Clang 16.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/100170
Approved by: https://github.com/ezyang