[inductor] Install guards on both cases of View.handle_negative_index (#103780)
This branch is not an optimization, it's a correctness issue so there should be
a guard installed on both sides of the branch. Otherwise we could have an
expression like `(s0 - s1)` that is initially positive, then becomes negative
with a new set of shapes and now references an invalid index.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/103780
Approved by: https://github.com/ezyang